job/owner-job/requirements.txt
redis
kubernetes
python-dotenv
setuptools
